Make sure only PDO parameter types are passed
… to PDO methods
