+++ src/Driver/PDO/Statement.php
@@ -2,13 +2,10 @@
 
 namespace Doctrine\DBAL\Driver\PDO;
 
-use Doctrine\DBAL\Driver\Exception as ExceptionInterface;
-use Doctrine\DBAL\Driver\Exception\UnknownParameterType;
 use Doctrine\DBAL\Driver\Result as ResultInterface;
 use Doctrine\DBAL\Driver\Statement as StatementInterface;
 use Doctrine\DBAL\ParameterType;
 use Doctrine\Deprecations\Deprecation;
-use PDO;
 use PDOException;
 use PDOStatement;
 
@@ -18,16 +15,6 @@ use function func_num_args;
 
 final class Statement implements StatementInterface
 {
-    private const PARAM_TYPE_MAP = [
-        ParameterType::NULL => PDO::PARAM_NULL,
-        ParameterType::INTEGER => PDO::PARAM_INT,
-        ParameterType::STRING => PDO::PARAM_STR,
-        ParameterType::ASCII => PDO::PARAM_STR,
-        ParameterType::BINARY => PDO::PARAM_LOB,
-        ParameterType::LARGE_OBJECT => PDO::PARAM_LOB,
-        ParameterType::BOOLEAN => PDO::PARAM_BOOL,
-    ];
-
     private PDOStatement $stmt;
 
     /** @internal The statement can be only instantiated by its driver connection. */
@@ -50,10 +37,10 @@ final class Statement implements StatementInterface
             );
         }
 
-        $type = $this->convertParamType($type);
+        $pdoType = ParameterTypeMap::convertParamType($type);
 
         try {
-            return $this->stmt->bindValue($param, $value, $type);
+            return $this->stmt->bindValue($param, $value, $pdoType);
         } catch (PDOException $exception) {
             throw Exception::new($exception);
         }
@@ -101,13 +88,13 @@ final class Statement implements StatementInterface
             );
         }
 
-        $type = $this->convertParamType($type);
+        $pdoType = ParameterTypeMap::convertParamType($type);
 
         try {
             return $this->stmt->bindParam(
                 $param,
                 $variable,
-                $type,
+                $pdoType,
                 $length ?? 0,
                 ...array_slice(func_get_args(), 4),
             );
@@ -137,21 +124,5 @@ final class Statement implements StatementInterface
         }
 
         return new Result($this->stmt);
-    }
-
-    /**
-     * Converts DBAL parameter type to PDO parameter type
-     *
-     * @param int $type Parameter type
-     *
-     * @throws ExceptionInterface
-     */
-    private function convertParamType(int $type): int
-    {
-        if (! isset(self::PARAM_TYPE_MAP[$type])) {
-            throw UnknownParameterType::new($type);
-        }
-
-        return self::PARAM_TYPE_MAP[$type];
     }
 }
